requirements.txt
# AtlasLeaf - Dependências
# =========================

# Deep Learning
torch==2.9.1
torchvision==0.24.1
torchaudio==2.9.1

# ONNX (Export e Inferência)
onnx==1.20.0
onnxruntime==1.23.2
onnxscript

# Interface Web
streamlit
pillow==12.1.0

# Utilitários
numpy==2.4.0
tqdm  # Progress bars

# Machine Learning (v3.0)
scikit-learn  # Métricas e splits
seaborn  # Visualizações
matplotlib  # Gráficos
